Task-46769 : Image disappearing after a reset server on notes (#1387)
* Task-46769 : Image disappearing after a reset server on notes
