+++ ecms-social-integration/pom.xml
@@ -11,7 +11,7 @@
   <name>eXo PLF:: ECMS - Social Activities</name>
   <description>ECMS with Social activity</description>
   <properties>
-    <exo.test.coverage.ratio>0.06</exo.test.coverage.ratio>
+    <exo.test.coverage.ratio>0.05</exo.test.coverage.ratio>
   </properties>
   <dependencies>
     <dependency>
